% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add two numbers}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{First input number to be added}

\item{y}{Second input number to be added, with default value equals 10c}
}
\value{
A number
}
\description{
Add two numbers
}
\examples{
x <- 6
y <- 4
my_add(x,y)
}
