% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string to vector}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vector within one string}

\item{split}{the delimiter of the character vector x}
}
\value{
A character vector
}
\description{
Split a string to vector
}
\examples{
x <- "i,am,beautiful"
strsplit1(x, split = ",")
}
