% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Adds 2 numbers}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{Number in numeric form}

\item{y}{Number in numeric form; defaults to 10.}
}
\value{
a number
}
\description{
Adds 2 numbers
}
\examples{
my_add(1, 2)
my_add(1)
}
